+++ mrv2/lib/mrvGL/mrvGLViewportDraw.cpp
@@ -23,6 +23,11 @@
 
 #include "mrViewer.h"
 
+namespace
+{
+    const unsigned kFPSAverageFrames = 10;
+}
+
 namespace mrv
 {
     void Viewport::_drawAnaglyph(int left, int right) const noexcept
@@ -910,34 +915,41 @@ namespace mrv
             if (player->playback() != timeline::Playback::Stop &&
                 (p.actionMode != ActionMode::kScrub || p.lastEvent != FL_DRAG))
             {
+                // Calculate skipped frames
                 int64_t frame_diff = (time.value() - p.lastTime.value());
                 int64_t absdiff = std::abs(frame_diff);
                 if (absdiff > 1 && absdiff < 60)
                     p.skippedFrames += absdiff - 1;
-                ++p.frameCount;
 
                 // Calculate elapsed time
-                auto endTime = std::chrono::high_resolution_clock::now();
-                auto elapsedTime =
-                    std::chrono::duration_cast<std::chrono::milliseconds>(
-                        endTime - p.startTime)
+                auto currentTime = std::chrono::high_resolution_clock::now();
+                auto frameTime =
+                    std::chrono::duration<double>(currentTime - p.startTime)
                         .count();
+                p.frameTimes.push_back(frameTime);
 
-                // Update FPS every second (adjust as needed)
-                if (elapsedTime >= 1000)
+                // Average FPS over kFPSAverageFrames
+                if (p.frameTimes.size() >= kFPSAverageFrames)
                 {
-                    p.actualFPS = p.frameCount / (elapsedTime / 1000.0);
-
-                    // Reset variables for the next second
-                    p.frameCount = 0;
-                    p.startTime = std::chrono::high_resolution_clock::now();
+                    p.frameTimes.pop_front();
                 }
 
+                // Calculate average frame time
+                double averageFrameTime = 0.0;
+                for (const double time : p.frameTimes)
+                {
+                    averageFrameTime += time;
+                }
+                averageFrameTime /= p.frameTimes.size();
+
+                const double fps = 1.0 / averageFrameTime;
+
                 snprintf(
-                    buf, 512, "SF: %" PRIu64 " FPS: %.3f/%.3f", p.skippedFrames,
-                    p.actualFPS, player->speed());
+                    buf, 512, "SF: %" PRIu64 " FPS: %.2f/%.3f", p.skippedFrames,
+                    fps, player->speed());
 
                 tmp += buf;
+                p.startTime = std::chrono::high_resolution_clock::now();
             }
         }